+++ CHANGELOG.md
@@ -5,12 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Changed
-- switch from github to repo for imports
 
+## [0.13.0] - July 12, 2021
 ### Added
 - stats package for info struct
 - format package for formatting text and defining flags
+- CPU MHz and tempurature data blocks
+
+### Changed
+- switch from github to custom repo for imports.
+- Battery update time from every 5 minutes to every minute.
 
 ## [0.12.0] - January 12, 2020